Update README.md
Need to fix login and register readme because currently not able to actually post successfully.

Am I missing endpoints?

Do we want front end to be able to retrieve a single user profile?

Need to fix error message for what happens if no students are registered and you try to get registered students for class

Do we have an endpoint for registering for a class? Do we need one?

Feel free to edit anything!